requirements.txt
elasticsearch
elasticsearch_curator
requests
